feat: DPU version can now decode Huffman bitstream and do inverse DCT
Changed datatype of MCU buffer elements from int to short to save space
in MRAM.

+++ include/jpeg-common.h
@@ -159,7 +159,7 @@ typedef struct JpegDecompressor {
  * MCU has a 2D array. 3 indices for color channels (RGB) and 64 indices the 64 values in an 8x8 MCU
  */
 typedef struct MCU {
-  int buffer[3][64];
+  short buffer[3][64];
 } MCU;
 
 void jpeg_cpu_scale(uint64_t file_length, char *filename, char *buffer);
